feat 214 : 백엔드 코드 수정, 마무리 2(#5)

+++ backend/src/main/java/com/mysite/knitly/domain/mypage/dto/FavoriteProductItem.java
@@ -10,7 +10,7 @@ public record FavoriteProductItem(
         Long productId,
         String productTitle,
         String thumbnailUrl,
-        BigDecimal price,
+        Double price,
         Double averageRating,
         LocalDate likedAt
 ) {}
